Port XTRX utils to vendored litepcie_util.
