--- software/app/.gitignore
@@ -1,1 +1,0 @@
-litex_xtrx_util
